Create InputManager
The preliminary InputManager class. This has only been half implemented. Still requires proper setting up of multiple binding schemes.
One of the things I'll eventually change is instead of having 1 file per scheme, 1 file with all the schemes in it would be better. It's not difficult to change it to that but I wanted something up quick.